+++ lib/netutil/tcplistener.go
@@ -62,7 +62,7 @@ func EnableIPv6() {
 	*enableTCP6 = true
 }
 
-// TCP6Enabled returns true if dialing and listening for IPv4 TCP is enabled.
+// TCP6Enabled returns true if IPv6 is enabled for dialing and listening.
 func TCP6Enabled() bool {
 	return *enableTCP6
 }